ingestion/requirements.txt
praw==7.7.1
feedparser==6.0.11
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
httpx==0.26.0
